apple-llvm-split-commit: 2d943f5ac1d1cdd6cd84c06edb5ef2b5f2c3a84b
apple-llvm-split-dir: lldb/

+++ libcxx/test/std/numerics/numeric.ops/inclusive.scan/inclusive_scan_op.pass.cpp
@@ -17,12 +17,9 @@
 //                    BinaryOperation binary_op); // C++17
 
 #include <numeric>
-#include <algorithm>
-#include <cassert>
-#include <functional>
-#include <iostream>
-#include <iterator>
 #include <vector>
+#include <cassert>
+#include <iostream>
 
 #include "test_iterators.h"
 
@@ -61,21 +58,21 @@ test()
         }
 }
 
-size_t triangle(size_t n) { return n*(n+1)/2; }
+int triangle(int n) { return n*(n+1)/2; }
 
 //  Basic sanity
 void basic_tests()
 {
     {
-    std::vector<size_t> v(10);
+    std::vector<int> v(10);
     std::fill(v.begin(), v.end(), 3);
     std::inclusive_scan(v.begin(), v.end(), v.begin(), std::plus<>());
     for (size_t i = 0; i < v.size(); ++i)
-        assert(v[i] == (i+1) * 3);
+        assert(v[i] == (int)(i+1) * 3);
     }
 
     {
-    std::vector<size_t> v(10);
+    std::vector<int> v(10);
     std::iota(v.begin(), v.end(), 0);
     std::inclusive_scan(v.begin(), v.end(), v.begin(), std::plus<>());
     for (size_t i = 0; i < v.size(); ++i)
@@ -83,7 +80,7 @@ void basic_tests()
     }
 
     {
-    std::vector<size_t> v(10);
+    std::vector<int> v(10);
     std::iota(v.begin(), v.end(), 1);
     std::inclusive_scan(v.begin(), v.end(), v.begin(), std::plus<>());
     for (size_t i = 0; i < v.size(); ++i)
@@ -91,7 +88,7 @@ void basic_tests()
     }
 
     {
-    std::vector<size_t> v, res;
+    std::vector<int> v, res;
     std::inclusive_scan(v.begin(), v.end(), std::back_inserter(res), std::plus<>());
     assert(res.empty());
     }
